backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlmodel==0.0.21
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pydantic[email]==2.8.0
python-jose[cryptography]==3.3.0
httpx==0.27.0
google-generativeai==0.7.2
